requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlmodel==0.0.16
dnspython==2.6.1
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1